aws-neuron/torch-neuronx/nxd-inference-whisper/requirements.txt
# AWS Neuron SDK (システム提供の仮想環境を使用)
# neuronxcc>=2.22
# torch-neuronx
# neuronx-distributed-inference>=0.7.0

# Hugging Face Transformers
transformers>=4.40.0

# OpenAI Whisper
openai-whisper

# 音声処理
soundfile
librosa

# ユーティリティ
numpy
scipy
